+++ packages/astro/src/core/config/schemas/base.ts
@@ -11,6 +11,7 @@ import { z } from 'zod';
 import { localFontFamilySchema, remoteFontFamilySchema } from '../../../assets/fonts/config.js';
 import { EnvSchema } from '../../../env/schema.js';
 import type { AstroUserConfig, ViteUserConfig } from '../../../types/public/config.js';
+import { allowedDirectivesSchema, cspAlgorithmSchema, cspHashSchema } from '../../csp/config.js';
 
 // The below types are required boilerplate to workaround a Zod issue since v3.21.2. Since that version,
 // Zod's compiled TypeScript would "simplify" certain values to their base representation, causing references
@@ -67,7 +68,7 @@ export const ASTRO_CONFIG_DEFAULTS = {
 	image: {
 		endpoint: { entrypoint: undefined, route: '/_image' },
 		service: { entrypoint: 'astro/assets/services/sharp', config: {} },
-		experimentalDefaultStyles: true,
+		responsiveStyles: false,
 	},
 	devToolbar: {
 		enabled: true,
@@ -97,9 +98,10 @@ export const ASTRO_CONFIG_DEFAULTS = {
 	experimental: {
 		clientPrerender: false,
 		contentIntellisense: false,
-		responsiveImages: false,
 		headingIdCompat: false,
 		preserveScriptOrder: false,
+		liveContentCollections: false,
+		csp: false,
 	},
 } satisfies AstroUserConfig & { server: { open: boolean } };
 
@@ -271,13 +273,11 @@ export const AstroConfigSchema = z.object({
 					}),
 				)
 				.default([]),
-			experimentalLayout: z.enum(['constrained', 'fixed', 'full-width', 'none']).optional(),
-			experimentalObjectFit: z.string().optional(),
-			experimentalObjectPosition: z.string().optional(),
-			experimentalBreakpoints: z.array(z.number()).optional(),
-			experimentalDefaultStyles: z
-				.boolean()
-				.default(ASTRO_CONFIG_DEFAULTS.image.experimentalDefaultStyles),
+			layout: z.enum(['constrained', 'fixed', 'full-width', 'none']).optional(),
+			objectFit: z.string().optional(),
+			objectPosition: z.string().optional(),
+			breakpoints: z.array(z.number()).optional(),
+			responsiveStyles: z.boolean().default(ASTRO_CONFIG_DEFAULTS.image.responsiveStyles),
 		})
 		.default(ASTRO_CONFIG_DEFAULTS.image),
 	devToolbar: z
@@ -432,7 +432,7 @@ export const AstroConfigSchema = z.object({
 		.default(ASTRO_CONFIG_DEFAULTS.env),
 	session: z
 		.object({
-			driver: z.string(),
+			driver: z.string().optional(),
 			options: z.record(z.any()).optional(),
 			cookie: z
 				.object({
@@ -464,10 +464,6 @@ export const AstroConfigSchema = z.object({
 				.boolean()
 				.optional()
 				.default(ASTRO_CONFIG_DEFAULTS.experimental.contentIntellisense),
-			responsiveImages: z
-				.boolean()
-				.optional()
-				.default(ASTRO_CONFIG_DEFAULTS.experimental.responsiveImages),
 			headingIdCompat: z
 				.boolean()
 				.optional()
@@ -477,6 +473,33 @@ export const AstroConfigSchema = z.object({
 				.optional()
 				.default(ASTRO_CONFIG_DEFAULTS.experimental.preserveScriptOrder),
 			fonts: z.array(z.union([localFontFamilySchema, remoteFontFamilySchema])).optional(),
+			liveContentCollections: z
+				.boolean()
+				.optional()
+				.default(ASTRO_CONFIG_DEFAULTS.experimental.liveContentCollections),
+			csp: z
+				.union([
+					z.boolean().optional().default(ASTRO_CONFIG_DEFAULTS.experimental.csp),
+					z.object({
+						algorithm: cspAlgorithmSchema,
+						directives: z.array(allowedDirectivesSchema).optional(),
+						styleDirective: z
+							.object({
+								resources: z.array(z.string()).optional(),
+								hashes: z.array(cspHashSchema).optional(),
+							})
+							.optional(),
+						scriptDirective: z
+							.object({
+								resources: z.array(z.string()).optional(),
+								hashes: z.array(cspHashSchema).optional(),
+								strictDynamic: z.boolean().optional(),
+							})
+							.optional(),
+					}),
+				])
+				.optional()
+				.default(ASTRO_CONFIG_DEFAULTS.experimental.csp),
 		})
 		.strict(
 			`Invalid or outdated experimental feature.\nCheck for incorrect spelling or outdated Astro version.\nSee https://docs.astro.build/en/reference/experimental-flags/ for a list of all current experiments.`,