+++ packages/astro/src/core/errors/errors-data.ts
@@ -1382,6 +1382,19 @@ export const FontFamilyNotFound = {
 
 /**
  * @docs
+ * @description
+ * The CSP feature isn't enabled
+ * @message
+ * The `experimental.csp` configuration isn't enabled.
+ */
+export const CspNotEnabled = {
+	name: 'CspNotEnabled',
+	title: "CSP feature isn't enabled",
+	message: "The `experimental.csp` configuration isn't enabled.",
+} satisfies ErrorData;
+
+/**
+ * @docs
  * @kind heading
  * @name CSS Errors
  */
@@ -1630,8 +1643,9 @@ export const InvalidContentEntryDataError = {
 	title: 'Content entry data does not match schema.',
 	message(collection: string, entryId: string, error: ZodError) {
 		return [
-			`**${String(collection)} → ${String(entryId)}** data does not match collection schema.`,
-			...error.errors.map((zodError) => zodError.message),
+			`**${String(collection)} → ${String(entryId)}** data does not match collection schema.\n`,
+			...error.errors.map((zodError) => `  **${zodError.path.join('.')}**: ${zodError.message}`),
+			'',
 		].join('\n');
 	},
 	hint: 'See https://docs.astro.build/en/guides/content-collections/ for more information on content schemas.',
@@ -1681,11 +1695,31 @@ export const ContentEntryDataError = {
 	title: 'Content entry data does not match schema.',
 	message(collection: string, entryId: string, error: ZodError) {
 		return [
-			`**${String(collection)} → ${String(entryId)}** data does not match collection schema.`,
-			...error.errors.map((zodError) => zodError.message),
+			`**${String(collection)} → ${String(entryId)}** data does not match collection schema.\n`,
+			...error.errors.map((zodError) => `  **${zodError.path.join('.')}**: ${zodError.message}`),
+			'',
 		].join('\n');
 	},
 	hint: 'See https://docs.astro.build/en/guides/content-collections/ for more information on content schemas.',
+} satisfies ErrorData;
+
+/**
+ * @docs
+ * @message
+ * **Example error message:**<br/>
+ * The schema cannot be a function for live collections. Please use a schema object instead. Check your collection definitions in your live content config file.
+ * @description
+ * Error in live content config.
+ * @see
+ * - [Experimental live content](https://docs.astro.build/en/reference/experimental-flags/live-content-collections/)
+ */
+
+export const LiveContentConfigError = {
+	name: 'LiveContentConfigError',
+	title: 'Error in live content config.',
+	message: (error: string, filename?: string) =>
+		`${error} Check your collection definitions in ${filename ?? 'your live content config file'}.`,
+	hint: 'See https://docs.astro.build/en/reference/experimental-flags/live-content-collections/ for more information on live content collections.',
 } satisfies ErrorData;
 
 /**